fix(L-05, create strike): fixes types in create event and strike as uint128 arg

+++ contracts/interfaces/engine/IPrimitiveEngineActions.sol
@@ -23,7 +23,7 @@ interface IPrimitiveEngineActions {
     /// delRisky            Total amount of risky tokens provided to reserves
     /// delStable           Total amount of stable tokens provided to reserves
     function create(
-        uint256 strike,
+        uint128 strike,
         uint32 sigma,
         uint32 maturity,
         uint32 gamma,
